Fixed bugs related to asynchronous calls made to server.
The main issue here was that the networking calls to the REST api worked fully but we have methods which require the return values from these REST api calls. Since the code which comes after these rest api calls keeps executing, their values were always null due to several milliseconds delay which comes from accessing the server. We had to turn some of the asynchronous calls to the api into synchronous ones to avoid this issue.