Changed the way user information is stored/accessed from the user class for better encapsulation.
Instead of accessing/setting each piece of info through the get/set methods, I have introduced a method which recieves a JsonObject(The response from server) and sets the fields of the class according to the info inside of it. Additionally, there is an object which returns a copy of that json object for display/editing of info which can then be sent to the server easily.